Ignoring missing foreign keys to ignored tables
This change prevents foreign key dependencies on ignored tables from being
generated by dbm-gorm-diff

The database-migration plug-in ignores missing foreign keys when the table
that they apply to is ignored, but did not when the table that is
referenced is ignored. These constraints, whilst not being directly
applied to the ignored tables, still depend on them.